refactor: moved games to minigame folder
